packages/api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "api"
description = "Property valuations API service."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "Anchor Point Risk", email = "henry@anchorpointrisk.co.za" }]
dynamic = ["version"]
dependencies = [
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.29,<1",
    "asyncpg>=0.29,<1",
    "pydantic>=2.6,<3",
    "pydantic-settings>=2,<3",
    "pyjwt[crypto]>=2.8,<3",
    "httpx>=0.27,<1",
    # v3+ changed import path to pythonjsonlogger.json.JsonFormatter; upgrade
    # requires updating api/logging.py to match before bumping this cap.
    "python-json-logger>=2.0,<3",
    "valuation-engine>=0.1,<0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-pyjwt",
]

[tool.uv.sources]
valuation-engine = { path = "../valuation_engine", editable = true }

[tool.hatch.version]
path = "src/api/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/api"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RET"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
files = ["src/api"]

[[tool.mypy.overrides]]
module = ["asyncpg", "asyncpg.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["valuation_engine", "valuation_engine.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["api", "api.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
asyncio_mode = "auto"
markers = [
    "integration: requires a running Supabase CLI stack (slow)",
]
